+++ src/radio/radio.ts
@@ -11,7 +11,12 @@ const iconDefault = {
 };
 @wxComponent()
 export default class Radio extends SuperComponent {
-  externalClasses = [`${prefix}-class`, `${prefix}-class-label`, `${prefix}-class-icon`, `${prefix}-class-content`];
+  externalClasses = [
+    `${prefix}-class`,
+    `${prefix}-class-label`,
+    `${prefix}-class-icon`,
+    `${prefix}-class-content`,
+  ];
 
   relations = {
     '../radio-group/radio-group': {
@@ -25,11 +30,26 @@ export default class Radio extends SuperComponent {
 
   lifetimes = {
     attached() {
-      this.handleInitStatus();
+      this.initStatus();
     },
   };
 
   properties = Props;
+
+  controlledProps = [
+    {
+      key: 'checked',
+      event: 'change',
+    },
+  ];
+
+  observers = {
+    checked(isChecked: Boolean) {
+      this.setData({
+        active: isChecked,
+      });
+    },
+  };
 
   data = {
     prefix,
@@ -42,59 +62,35 @@ export default class Radio extends SuperComponent {
   };
 
   methods = {
-    onChange(e) {
+    handleTap(e) {
       if (this.data.disabled) return;
+
       const { target } = e.currentTarget.dataset;
-      const { contentDisabled } = this.data;
-      if (target === 'text' && contentDisabled) {
-        return;
-      }
-      const { value, active, optionLinked } = this.data;
+
+      if (target === 'text' && this.data.contentDisabled) return;
+
+      const { value, active } = this.data;
       const [parent] = this.getRelationNodes('../radio-group/radio-group');
+
       if (parent) {
-        parent.updateValue({ name: value });
+        parent.updateValue(value);
       } else {
-        if (optionLinked) {
-          this.triggerEvent('toggleGroupSelect', { name: value });
-          return;
-        }
-        this.triggerEvent('change', !active);
-        this.toggle();
+        this._trigger('change', { checked: !active });
       }
     },
-    handleInitStatus() {
+    initStatus() {
       const { icon } = this.data;
       const isIdArr = Array.isArray(icon);
+
       this.setData({
         customIcon: isIdArr,
         iconVal: !isIdArr ? iconDefault[icon] : this.data.icon,
       });
-      if (!this.data.optionLinked) {
-        this.setData({
-          active: this.data.checked,
-        });
-      }
     },
-    toggle() {
-      const { active } = this.data;
-      this.setData({
-        active: !active,
-      });
-    },
-    changeActive(active: boolean) {
-      this.setData({
-        active,
-      });
-    },
+
     setDisabled(disabled: Boolean) {
       this.setData({
         disabled: this.data.disabled || disabled,
-      });
-    },
-    // 支持options
-    setOptionLinked(linked: Boolean) {
-      this.setData({
-        optionLinked: linked,
       });
     },
   };