+++ src/stepper/props.ts
@@ -51,6 +51,11 @@ const props: TdStepperProps = {
     optionalTypes: [Number],
     value: null,
   },
+  /** 值 - 非受控 */
+  defaultValue: {
+    type: null,
+    value: undefined,
+  },
 };
 
 export default props;