+++ src/tag/check-tag.ts
@@ -14,9 +14,22 @@ export default class CheckTag extends SuperComponent {
 
   properties = props;
 
+  controlledProps = [
+    {
+      key: 'checked',
+      event: 'change',
+    },
+  ];
+
   methods = {
     onClickClose(e: WechatMiniprogram.BaseEvent) {
       this.triggerEvent('close', e);
+    },
+
+    handleChange() {
+      if (this.data.disabled) return;
+
+      this._trigger('change', { checked: !this.data.checked });
     },
   };
 }